# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a file and folder explorer API",
        "title": "Remy Explorer API",
        "contact": {
            "name": "Remy Team",
            "email": "remystorage@yandex.ru"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "0.0.2"
    },
    "basePath": "/",
    "paths": {
        "/files": {
            "put": {
                "description": "Update the details of an existing file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Update a file",
                "parameters": [
                    {
                        "description": "Update File Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateFileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new file in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Create a new file",
                "parameters": [
                    {
                        "description": "Create File Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateFileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "description": "Retrieve a file's details by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get file by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetFileByIDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a file by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DeleteFileResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/folders": {
            "put": {
                "description": "Update the details of an existing folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Update a folder",
                "parameters": [
                    {
                        "description": "Update Folder Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateFolderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new folder in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Create a new folder",
                "parameters": [
                    {
                        "description": "Create Folder Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateFolderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/folders/{folderID}/files": {
            "get": {
                "description": "Retrieve a list of files in a specific folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get files by folder ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetFilesByFolderIDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/folders/{id}": {
            "get": {
                "description": "Retrieve a folder's details by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Get folder by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetFolderByIDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a folder by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Delete a folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DeleteFolderResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/folders/{parentID}/subfolders": {
            "get": {
                "description": "Retrieve a list of folders within a specific parent folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Get folders by parent ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parent Folder ID",
                        "name": "parentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.GetFoldersByParentIDResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schemas.CreateFileRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "folder_id": {
                    "description": "ID of the parent folder",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the file",
                    "type": "string"
                },
                "owner_id": {
                    "description": "ID of the owner",
                    "type": "string"
                },
                "path": {
                    "description": "Path where the file is stored",
                    "type": "string"
                },
                "size": {
                    "description": "Size of the file",
                    "type": "integer"
                },
                "type": {
                    "description": "Type of the file",
                    "type": "string"
                }
            }
        },
        "schemas.CreateFileResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the created file",
                    "type": "string"
                }
            }
        },
        "schemas.CreateFolderRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "Name of the folder",
                    "type": "string"
                },
                "owner_id": {
                    "description": "ID of the owner",
                    "type": "string"
                },
                "parent_id": {
                    "description": "ID of the parent folder",
                    "type": "string"
                }
            }
        },
        "schemas.CreateFolderResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the created folder",
                    "type": "string"
                }
            }
        },
        "schemas.DeleteFileResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "description": "Indicates whether the deletion was successful",
                    "type": "boolean"
                }
            }
        },
        "schemas.DeleteFolderResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "description": "Indicates whether the deletion was successful",
                    "type": "boolean"
                }
            }
        },
        "schemas.ErrorResponse": {
            "description": "Represents a standard error response for the API",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer"
                },
                "message": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
        "schemas.GetFileByIDResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp when the file was created",
                    "type": "string"
                },
                "folder_id": {
                    "description": "ID of the parent folder",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the file",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the file",
                    "type": "string"
                },
                "path": {
                    "description": "Path where the file is stored",
                    "type": "string"
                },
                "size": {
                    "description": "Size of the file",
                    "type": "integer"
                },
                "tags": {
                    "description": "Tags associated with the file",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "Type of the file",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Timestamp when the file was last updated",
                    "type": "string"
                }
            }
        },
        "schemas.GetFilesByFolderIDResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "description": "List of files",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.ShortFileInfo"
                    }
                },
                "length": {
                    "description": "Number of files",
                    "type": "integer"
                }
            }
        },
        "schemas.GetFolderByIDResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp when the folder was created",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the folder",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the folder",
                    "type": "string"
                },
                "owner_id": {
                    "description": "ID of the owner",
                    "type": "string"
                },
                "parent_id": {
                    "description": "ID of the parent folder",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Timestamp when the folder was last updated",
                    "type": "string"
                }
            }
        },
        "schemas.GetFoldersByParentIDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the folder",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the folder",
                    "type": "string"
                }
            }
        },
        "schemas.ShortFileInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the file",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the file",
                    "type": "string"
                },
                "type": {
                    "description": "Type of the file",
                    "type": "string"
                }
            }
        },
        "schemas.UpdateFileRequest": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "folder_id": {
                    "description": "New parent folder ID",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the file to update",
                    "type": "string"
                },
                "name": {
                    "description": "New name of the file",
                    "type": "string"
                }
            }
        },
        "schemas.UpdateFileResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "description": "Indicates whether the update was successful",
                    "type": "boolean"
                }
            }
        },
        "schemas.UpdateFolderRequest": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "description": "ID of the folder to update",
                    "type": "string"
                },
                "name": {
                    "description": "New name of the folder",
                    "type": "string"
                },
                "parent_id": {
                    "description": "New parent folder ID",
                    "type": "string"
                }
            }
        },
        "schemas.UpdateFolderResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "description": "Indicates whether the update was successful",
                    "type": "boolean"
                }
            }
        }
    }
}